requirements.txt
# Brain Tumor Detection Project - Requirements
# Install: pip install -r requirements.txt

# Core ML Libraries
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.23.0
pandas>=2.0.0
scikit-learn>=1.2.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.5.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Model Evaluation
scikit-plot>=0.3.7

# Web Application
streamlit>=1.22.0
flask>=2.3.0

# Utilities
tqdm>=4.65.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.22.0

# Testing
pytest>=7.3.0

# Logging
tensorboard>=2.12.0

# Optional: GPU Support (uncomment if using CUDA)
# tensorflow-gpu>=2.12.0
